+++ test/OpenTelemetry.Instrumentation.EntityFrameworkCore.Tests/EntityFrameworkDiagnosticListenerTests.cs
@@ -423,6 +423,7 @@ public class EntityFrameworkDiagnosticListenerTests : IDisposable
         // TBD: SqlLite not setting the DataSource so it doesn't get set.
         Assert.DoesNotContain(activity.Tags, t => t.Key == "peer.service");
         Assert.DoesNotContain(activity.Tags, t => t.Key == "server.address");
+        Assert.DoesNotContain(activity.Tags, t => t.Key == "server.port");
 
         if (!emitNewAttributes)
         {