+++ src/OpenTelemetry.Instrumentation.EntityFrameworkCore/CHANGELOG.md
@@ -8,6 +8,10 @@
 * Fix `db.system.name` values to follow new database semantic conventions when opted
   into using the `OTEL_SEMCONV_STABILITY_OPT_IN` environment variable.
   ([#3004](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3004))
+* Add the `server.port` resource attribute when following the new database semantic
+  conventions when opted into using the `OTEL_SEMCONV_STABILITY_OPT_IN` environment
+  variable.
+  ([#3011](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3011))
 
 ## 1.12.0-beta.2